Cadastro de produtos com HTMX

+++ produto/models.py
@@ -5,8 +5,8 @@ from django.db import models
 
 class Produto(models.Model):
     nome = models.CharField(max_length=99)
-    quantidade = models.PositiveIntegerField()
-    preco = models.PositiveIntegerField()
+    quantidade = models.IntegerField()
+    preco = models.DecimalField(max_digits=9, decimal_places=2)
 
     def __str__(self):
         return self.nome
